requirements.txt
aiofiles==0.6.0
aiohttp==3.7.4.post0
aiosqlite==0.17.0
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
asyncio-extras==1.3.2
asyncpg==0.23.0
asyncpraw==7.3.1
asyncprawcore==2.2.1
attrs==21.2.0
braceexpand==0.1.7
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
discord.py==1.7.2
emojis==0.6.0
google-api-core==1.28.0
google-api-python-client==2.7.0
google-auth==1.30.1
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
humanize==3.6.0
idna==2.10
import-expression==1.1.4
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
jishaku==1.20.0.220
MarkupSafe==2.0.1
multidict==5.1.0
packaging==20.9
Pillow==8.2.0
pipenv==2018.11.26
protobuf==3.17.1
psutil==5.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dotenv==0.17.1
pytz==2021.1
requests==2.25.1
rsa==4.7.2
six==1.16.0
tabulate==0.8.9
typing-extensions==3.10.0.0
update-checker==0.18.0
uritemplate==3.0.1
urllib3==1.26.5
Werkzeug==2.0.1
yarl==1.6.3
zipp==3.5.0